Move collective_assert_tensor to distributed utils
